requirements.txt
selenium==4.39.0
beautifulsoup4==4.13.4
python-dotenv==1.1.0